requirements.txt
huggingface_hub>=0.23.0
Pillow>=10.0.0
requests>=2.31.0
